rewrite shared store as Svelte5
